backend-master/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==5.0.2
django-allauth==0.50.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-location-field==2.7.3
django-rest-auth==0.9.5
django-rest-durin==1.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.12.6
hiredis==2.3.2
humanize==4.9.0
idna==3.6
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
